+++ tests/common/mod.rs
@@ -14,18 +14,18 @@ use rand::rngs::ThreadRng;
 use rand::seq::SliceRandom;
 use rand::thread_rng;
 use serde_json::Value;
-use acars_vdlm2_parser::{AcarsVdlm2Message, DecodeMessage, MessageResult};
+use acars_vdlm2_parser::{AcarsVdlm2Message, MessageResult};
 use acars_vdlm2_parser::acars::{AcarsMessage, NewAcarsMessage};
 use acars_vdlm2_parser::vdlm2::{NewVdlm2Message, Vdlm2Message};
 
 /// Enum for indicating test data type.
-pub(crate) enum MessageType {
+pub enum MessageType {
     Acars,
     Vdlm2,
     All,
 }
 
-pub(crate) enum SpeedTestType {
+pub enum SpeedTestType {
     IteratingRoundsLibrary,
     LargeQueueLibrary,
     IteratingRoundsValue,
@@ -43,7 +43,7 @@ impl fmt::Display for SpeedTestType {
     }
 }
 
-pub(crate) enum StopwatchType {
+pub enum StopwatchType {
     AllDeser,
     AllSer,
     LargeQueueSer,
@@ -51,27 +51,32 @@ pub(crate) enum StopwatchType {
     TotalRun
 }
 
-pub(crate) enum StatType {
+pub enum StatType {
     AllDeser,
     AllSer
 }
 
 /// Struct for storing test information for the tests that just display error information.
-pub(crate) struct TestFile {
-    pub(crate) name: String,
-    pub(crate) contents: Vec<String>,
+pub struct TestFile {
+    pub name: String,
+    pub contents: Vec<String>,
 }
 
-pub(crate) struct Stopwatch {
-    pub(crate) start_time: Option<DateTime<Utc>>,
-    pub(crate) stop_time: Option<DateTime<Utc>>,
-    pub(crate) duration_ms: i64,
-    pub(crate) duration_ns: i64,
-    pub(crate) stopwatch_type: StopwatchType
+
+/// Struct for storing the start, end time and durations for doing elapsed time measurement.
+pub struct Stopwatch {
+    pub start_time: Option<DateTime<Utc>>,
+    pub stop_time: Option<DateTime<Utc>>,
+    pub duration_ms: i64,
+    pub duration_ns: i64,
+    pub stopwatch_type: StopwatchType
 }
 
 impl Stopwatch {
-    pub(crate) fn start(stopwatch_type: StopwatchType) -> Self {
+    /// Set the start DateTime for when the call is made and store it.
+    ///
+    /// Returns an instance of itself
+    pub fn start(stopwatch_type: StopwatchType) -> Self {
         Self {
             start_time: Some(Utc::now()),
             stop_time: None,
@@ -80,7 +85,10 @@ impl Stopwatch {
             stopwatch_type
         }
     }
-    pub(crate) fn stop(&mut self) {
+    /// Sets the stop DateTime for when the call is made and stores it.
+    ///
+    /// Will also calculate the duration in milliseconds and nanoseconds and store them in two i64's
+    pub fn stop(&mut self) {
         self.stop_time = Some(Utc::now());
         if let (Some(stop), Some(start)) = (self.stop_time, self.start_time) {
             let duration: chrono::Duration = stop - start;
@@ -92,33 +100,32 @@ impl Stopwatch {
     }
 }
 
-// #[allow(dead_code)]
 #[derive(Debug, Clone, Default)]
-pub(crate) struct RunDurations {
-    pub(crate) run_processed_items: usize,
-    pub(crate) total_test_runs: i64,
-    pub(crate) test_runs: Vec<TestRun>,
-    pub(crate) large_queue_ser_ms: i64,
-    pub(crate) large_queue_ser_ns: i64,
-    pub(crate) large_queue_deser_ms: i64,
-    pub(crate) large_queue_deser_ns: i64,
-    pub(crate) total_run_ms: i64,
-    pub(crate) total_run_ns: i64
+pub struct RunDurations {
+    pub run_processed_items: usize,
+    pub total_test_runs: i64,
+    pub test_runs: Vec<TestRun>,
+    pub large_queue_ser_ms: i64,
+    pub large_queue_ser_ns: i64,
+    pub large_queue_deser_ms: i64,
+    pub large_queue_deser_ns: i64,
+    pub total_run_ms: i64,
+    pub total_run_ns: i64
     
 }
 
 #[derive(Debug, Clone, Default)]
-pub(crate) struct TestRun {
-    pub(crate) run_number: i64,
-    pub(crate) run_items: usize,
-    pub(crate) deser_ms: i64,
-    pub(crate) deser_ns: i64,
-    pub(crate) ser_ms: i64,
-    pub(crate) ser_ns: i64
+pub struct TestRun {
+    pub run_number: i64,
+    pub run_items: usize,
+    pub deser_ms: i64,
+    pub deser_ns: i64,
+    pub ser_ms: i64,
+    pub ser_ns: i64
 }
 
 impl TestRun {
-    pub(crate) fn new(run_number: &i64) -> Self {
+    pub fn new(run_number: &i64) -> Self {
         Self {
             run_number: *run_number,
             run_items: usize::default(),
@@ -129,7 +136,7 @@ impl TestRun {
         }
     }
     
-    pub(crate) fn update_run_durations(&mut self, stopwatch: &Stopwatch) {
+    pub fn update_run_durations(&mut self, stopwatch: &Stopwatch) {
         match stopwatch.stopwatch_type {
             StopwatchType::AllDeser => {
                 self.deser_ms = stopwatch.duration_ms;
@@ -146,11 +153,8 @@ impl TestRun {
     }
 }
 
-
-
-// #[allow(dead_code)]
 impl RunDurations {
-    pub(crate) fn new() -> Self {
+    pub fn new() -> Self {
         Self {
             run_processed_items: usize::default(),
             total_test_runs: i64::default(),
@@ -163,7 +167,7 @@ impl RunDurations {
             total_run_ns: i64::default()
         }
     }
-    pub(crate) fn update_run_durations(&mut self, stopwatch: &Stopwatch) {
+    pub fn update_run_durations(&mut self, stopwatch: &Stopwatch) {
         match stopwatch.stopwatch_type {
             StopwatchType::AllDeser => {}
             StopwatchType::AllSer => {}
@@ -183,15 +187,15 @@ impl RunDurations {
     }
 }
 
-pub(crate) struct SpeedTestComparisons {
-    pub(crate) test_one_type: SpeedTestType,
-    pub(crate) test_one_results: RunDurations,
-    pub(crate) test_two_type: SpeedTestType,
-    pub(crate) test_two_results: RunDurations
+pub struct SpeedTestComparisons {
+    pub test_one_type: SpeedTestType,
+    pub test_one_results: RunDurations,
+    pub test_two_type: SpeedTestType,
+    pub test_two_results: RunDurations
 }
 
 impl SpeedTestComparisons {
-    pub(crate) fn compare_rounds(self) {
+    pub fn compare_rounds(self) {
         let mut comparison_table: Table = Table::new();
         let mut test_one: RunDurations = self.test_one_results;
         let mut test_two: RunDurations = self.test_two_results;
@@ -282,7 +286,7 @@ impl SpeedTestComparisons {
         ]);
         comparison_table.printstd();
     }
-    pub(crate) fn compare_large_queue(self) {
+    pub fn compare_large_queue(self) {
         let mut comparison_table: Table = Table::new();
         let test_one: RunDurations = self.test_one_results;
         let test_two: RunDurations = self.test_two_results;
@@ -309,21 +313,21 @@ impl SpeedTestComparisons {
     }
 }
 
-pub(crate) struct SpeedTestDurations {
-    pub(crate) shortest_duration_ms: Option<i64>,
-    pub(crate) shortest_duration_ns: Option<i64>,
-    pub(crate) shortest_run_number: i64,
-    pub(crate) shortest_run_items: usize,
-    pub(crate) longest_duration_ms: Option<i64>,
-    pub(crate) longest_duration_ns: Option<i64>,
-    pub(crate) longest_run_number: i64,
-    pub(crate) longest_run_items: usize,
-    pub(crate) average_duration_ms: i64,
-    pub(crate) average_duration_ns: i64,
-    pub(crate) average_duration_items: usize,
+pub struct SpeedTestDurations {
+    pub shortest_duration_ms: Option<i64>,
+    pub shortest_duration_ns: Option<i64>,
+    pub shortest_run_number: i64,
+    pub shortest_run_items: usize,
+    pub longest_duration_ms: Option<i64>,
+    pub longest_duration_ns: Option<i64>,
+    pub longest_run_number: i64,
+    pub longest_run_items: usize,
+    pub average_duration_ms: i64,
+    pub average_duration_ns: i64,
+    pub average_duration_items: usize,
 }
 
-pub(crate) trait BuildSpeedTestDurations {
+pub trait BuildSpeedTestDurations {
     fn get_run_stats(&mut self, stat_type: StatType) -> SpeedTestDurations;
 }
 
@@ -409,7 +413,7 @@ impl BuildSpeedTestDurations for Vec<TestRun> {
 /// Trait for appending data.
 ///
 /// Using a trait to allow for implementation against `Vec<TestFile>`.
-pub(crate) trait AppendData {
+pub trait AppendData {
     fn append_data(&mut self, file: GlobResult) -> Result<(), Box<dyn Error>>;
 }
 
@@ -456,14 +460,14 @@ impl AppendData for Vec<TestFile> {
 /// Assistance function for tests to read a file, and break it up per line to a `Vec<String>`.
 ///
 /// This allows for tests to iterate through and test each line individually.
-pub(crate) fn read_test_file(filepath: impl AsRef<Path>) -> io::Result<Vec<String>> {
+pub fn read_test_file(filepath: impl AsRef<Path>) -> io::Result<Vec<String>> {
     BufReader::new(File::open(filepath)?).lines().collect()
 }
 
 /// Assistane function to combine contents of test files into a `Vec<String>`.
 ///
 /// This is used for combining the contents of multiple files into a single `Vec<String>` for testing.
-pub(crate) fn combine_found_files(
+pub fn combine_found_files(
     find_files: Result<Paths, PatternError>,
 ) -> Result<Vec<String>, Box<dyn Error>> {
     match find_files {
@@ -482,7 +486,7 @@ pub(crate) fn combine_found_files(
 }
 
 /// Assistance function for building a `Vec<TestFile>` for use with the tests that show parsing output.
-pub(crate) fn load_found_files(
+pub fn load_found_files(
     find_files: Result<Paths, PatternError>,
 ) -> Result<Vec<TestFile>, Box<dyn Error>> {
     match find_files {
@@ -501,7 +505,7 @@ pub(crate) fn load_found_files(
 }
 
 /// Assistance function for appending file contents.
-pub(crate) fn append_lines(
+pub fn append_lines(
     file: GlobResult,
     data: &mut Vec<String>,
 ) -> Result<(), Box<dyn Error>> {
@@ -523,7 +527,7 @@ pub(crate) fn append_lines(
 }
 
 /// Assistance function that combines contents of message type test files.
-pub(crate) fn combine_files_of_message_type(
+pub fn combine_files_of_message_type(
     message_type: MessageType,
 ) -> Result<Vec<String>, Box<dyn Error>> {
     match message_type {
@@ -543,7 +547,7 @@ pub(crate) fn combine_files_of_message_type(
 }
 
 /// Assistance function that loads contents of individual message type test files and returns them separately instead of combined.
-pub(crate) fn load_files_of_message_type(
+pub fn load_files_of_message_type(
     message_type: MessageType,
 ) -> Result<Vec<TestFile>, Box<dyn Error>> {
     match message_type {
@@ -563,7 +567,7 @@ pub(crate) fn load_files_of_message_type(
 }
 
 /// Assistance function for processing the contents of a `&[String]` slice as vdlm2 messages.
-pub(crate) fn process_file_as_vdlm2(contents: &[String]) {
+pub fn process_file_as_vdlm2(contents: &[String]) {
     let contents: Vec<String> = contents.to_vec();
     let mut errors: Vec<String> = Vec::new();
     for (entry, line) in contents.iter().enumerate() {
@@ -590,7 +594,7 @@ pub(crate) fn process_file_as_vdlm2(contents: &[String]) {
 }
 
 /// Assistance function for processing the contents of a `&[String]` slice as acars messages.
-pub(crate) fn process_file_as_acars(contents: &[String]) {
+pub fn process_file_as_acars(contents: &[String]) {
     let contents: Vec<String> = contents.to_vec();
     let mut errors: Vec<String> = Vec::new();
     for (entry, line) in contents.iter().enumerate() {
@@ -617,7 +621,7 @@ pub(crate) fn process_file_as_acars(contents: &[String]) {
 }
 
 /// Assistance function to compare error message strings between Library result and serde `Value` result.
-pub(crate) fn compare_errors(
+pub fn compare_errors(
     error_1: Option<serde_json::Error>,
     error_2: Result<Value, serde_json::Error>,
     line: &str,
@@ -635,7 +639,7 @@ pub(crate) fn compare_errors(
     }
 }
 
-pub(crate) fn test_enum_serialisation(message: &AcarsVdlm2Message) {
+pub fn test_enum_serialisation(message: &AcarsVdlm2Message) {
     let encoded_string: MessageResult<String> = message.to_string();
     assert_eq!(
         encoded_string.as_ref().err().is_none(),
@@ -654,7 +658,7 @@ pub(crate) fn test_enum_serialisation(message: &AcarsVdlm2Message) {
     );
 }
 
-pub(crate) fn test_value_serialisation(message: &Value) {
+pub fn test_value_serialisation(message: &Value) {
     let encoded_string: MessageResult<String> = serde_json::to_string(&message);
     assert_eq!(
         encoded_string.as_ref().err().is_none(),
@@ -673,7 +677,7 @@ pub(crate) fn test_value_serialisation(message: &Value) {
     );
 }
 
-pub(crate) trait ContentDuplicator {
+pub trait ContentDuplicator {
     fn duplicate_contents(&self, rounds: &i64) -> Self;
 }
 
@@ -689,160 +693,5 @@ impl ContentDuplicator for Vec<String> {
             }
         }
         duplicated_contents
-    }
-}
-
-
-/// This test will ingest contents from the vdlm2 sample files as a message per line to a `Vec<String>`.
-/// It combines the two files together into a single `Vec<String>` for iterating through.
-/// Then it will cycle them into `Vec<Vdlm2Message>` and back to `String`.
-/// It validates that there are no errors going `String` -> `Vdlm2Message` and `Vdlm2Message` -> `String`.
-#[test]
-fn test_vdlm2_parsing() -> Result<(), Box<dyn Error>> {
-    let load_vdlm_messages: Result<Vec<String>, Box<dyn Error>> =
-        combine_files_of_message_type(MessageType::Vdlm2);
-    match load_vdlm_messages {
-        Err(load_failed) => Err(load_failed),
-        Ok(vdlm2_messages) => {
-            let mut valid_vdlm2_messages: Vec<Vdlm2Message> = Vec::new();
-            let mut failed_decodes: Vec<String> = Vec::new();
-            for line in vdlm2_messages {
-                let parse_line: MessageResult<Vdlm2Message> = line.to_vdlm2();
-                match parse_line {
-                    Err(_) => failed_decodes.push(line),
-                    Ok(valid_entry) => valid_vdlm2_messages.push(valid_entry),
-                }
-            }
-            for message in valid_vdlm2_messages {
-                let vdlm2_to_string: MessageResult<String> = message.to_string();
-                assert_eq!(vdlm2_to_string.as_ref().err().is_none(), true);
-            }
-            for line in failed_decodes {
-                let library_parse_error: Option<serde_json::Error> = line.to_vdlm2().err();
-                let serde_value_error: Result<Value, serde_json::Error> =
-                    serde_json::from_str(&line);
-                compare_errors(library_parse_error, serde_value_error, &line);
-            }
-            Ok(())
-        }
-    }
-}
-
-/// Test for displaying the per-item result for vdlm2 messages, helpful when diagnosing parsing issues.
-/// Marked as `#[ignore]` so it can be run separately as required.
-#[test]
-#[ignore]
-fn show_vdlm2_ingest() -> Result<(), Box<dyn Error>> {
-    println!("Showing vdlm2 ingest errors");
-    let load_vdlm2_files: Result<Vec<TestFile>, Box<dyn Error>> =
-        load_files_of_message_type(MessageType::Vdlm2);
-    match load_vdlm2_files {
-        Err(load_failed) => Err(load_failed),
-        Ok(vdlm2_files) => {
-            for file in vdlm2_files {
-                println!("Testing the contents from file: {}", file.name);
-                process_file_as_vdlm2(&file.contents);
-            }
-            Ok(())
-        }
-    }
-}
-
-/// This test will ingest contents from the acars sample files as a message per line to a `Vec<String>`.
-/// It combines the two files together into a single `Vec<String>` for iterating through.
-/// Then it will cycle them into `Vec<AcarsMessage>` and back to `String`.
-/// It validates that there are no errors going `String` -> `AcarsMessage` and `AcarsMessage` -> `String`.
-#[test]
-fn test_acars_parsing() -> Result<(), Box<dyn Error>> {
-    let load_acars_messages: Result<Vec<String>, Box<dyn Error>> =
-        combine_files_of_message_type(MessageType::Acars);
-    match load_acars_messages {
-        Err(load_failed) => Err(load_failed),
-        Ok(acars_messages) => {
-            let mut valid_acars_messages: Vec<AcarsMessage> = Vec::new();
-            let mut failed_decodes: Vec<String> = Vec::new();
-            for line in acars_messages {
-                let parse_line: MessageResult<AcarsMessage> = line.to_acars();
-                match parse_line {
-                    Err(_) => failed_decodes.push(line),
-                    Ok(acars_message) => valid_acars_messages.push(acars_message),
-                }
-            }
-            for message in valid_acars_messages {
-                let acars_to_string: MessageResult<String> = message.to_string();
-                assert_eq!(acars_to_string.as_ref().err().is_none(), true);
-                let acars_to_bytes: MessageResult<Vec<u8>> = message.to_bytes();
-                assert_eq!(acars_to_bytes.as_ref().err().is_none(), true);
-            }
-            for line in failed_decodes {
-                let library_parse_error: Option<serde_json::Error> = line.to_acars().err();
-                let serde_value_error: Result<Value, serde_json::Error> =
-                    serde_json::from_str(&line);
-                compare_errors(library_parse_error, serde_value_error, &line);
-            }
-            Ok(())
-        }
-    }
-}
-
-/// Test for displaying the per-item result for acars messages, helpful when diagnosing parsing issues.
-/// Marked as `#[ignore]` so it can be run separately as required.
-#[test]
-#[ignore]
-fn show_acars_ingest() -> Result<(), Box<dyn Error>> {
-    println!("Showing acars ingest errors");
-    let load_acars_files: Result<Vec<TestFile>, Box<dyn Error>> =
-        load_files_of_message_type(MessageType::Acars);
-    match load_acars_files {
-        Err(load_failed) => Err(load_failed),
-        Ok(acars_files) => {
-            for file in acars_files {
-                println!("Testing the contents from file: {}", file.name);
-                process_file_as_acars(&file.contents);
-            }
-            Ok(())
-        }
-    }
-}
-
-/// This test ingests the contents of all the Acars and Vdlm2 sample files into individual `Vec<String>`.
-/// Then it combines two pairs of known good into a single `Vec<String>` and randomises the ordering.
-/// Then it will cycle them into `Vec<AcarsVdlm2Message>` and back to `String`.
-/// It validates that there are no errors going `String` -> `AcarsVdlm2Message` and `AcarsVdlm2Message` -> `String`.
-/// It then combines two files containing known bad data into a single `Vec<String>` and randomises the ordering.
-/// It validates that it gets errors that it is expecting and the correct number of errors.
-#[test]
-fn test_determining_message() -> Result<(), Box<dyn Error>> {
-    let load_all_messages: Result<Vec<String>, Box<dyn Error>> =
-        combine_files_of_message_type(MessageType::All);
-    match load_all_messages {
-        Err(load_error) => Err(load_error),
-        Ok(mut all_messages) => {
-            let mut rng: ThreadRng = thread_rng();
-            let mut successfully_decoded_items: Vec<AcarsVdlm2Message> = Vec::new();
-            let mut failed_decodes: Vec<String> = Vec::new();
-            all_messages.shuffle(&mut rng);
-            for entry in all_messages {
-                let parsed_message: MessageResult<AcarsVdlm2Message> = entry.decode_message();
-                match parsed_message {
-                    Err(_) =>
-                        failed_decodes.push(entry),
-                    Ok(decoded_message) =>
-                        successfully_decoded_items.push(decoded_message),
-                }
-            }
-            successfully_decoded_items.shuffle(&mut rng);
-            for message in successfully_decoded_items {
-                test_enum_serialisation(&message);
-            }
-            for line in failed_decodes {
-                let library_parse_error: Option<serde_json::Error> =
-                    line.decode_message().err();
-                let serde_value_error: Result<Value, serde_json::Error> =
-                    serde_json::from_str(&line);
-                compare_errors(library_parse_error, serde_value_error, &line);
-            }
-            Ok(())
-        }
     }
 }
